+++ front/app/containers/App/ModalQueue/modals/CommunityMonitor/CommunityMonitorModal.test.tsx
