+++ front/app/containers/App/ModalQueue/modals/CommunityMonitor/events.ts
