+++ front/app/containers/App/ModalQueue/modals/CommunityMonitor/index.tsx
